test: mark validation tests with doc annotation
Tests that are meant to be validation tests (i.e. are supposed to be
compared against Obsidian, and used to ensure our implementation matches
that of Obsidian's) should have `@validation-test` in the multi-line
comment block at the top of the file.

This doesn't mean anything for the scripts, but it helps humans identify
what the purpose of tests are.